Introduced concept of callback functions
The idea is that we don't perform the filtering on each and every entry
but rather just when necessary.

For now, let's do it just once: the first time a trace entry is hit. We
perform the filtering and then return a callback function which either
adds an entry or does nothing, depending on the filter results.