Divided TraceEntry into TracePoint and TraceEntry
A TracePoint is a static object which has a fixed location in the
source code as well as a configuration. The configuration of a trace point
tells whether a trace point is actiev or not, whether hitting it should
yield backtraces and whether variable values should be logged when hitting
the trace point.

A Trace Entry represents a single visit to a TracePoint. A TraceEntry is
fixed in time and shortlived. It references a TracePoint and possibly also
contains additional information such as a backtrace or variable values.

Filtering is done on TracePoints, not on trace entries.