Folded Trace::addEntry() into Trace::visitTracePoint()
The new Trace::visitTracePoint() function contains most of the code of
the TRACE_ macros. This not only makes the footsprint of the macros
smaller, it also allows more optimizations. For instance, we now don't
generate backtarce at all in case the current output object of the
trace cannot be written to.

Such optimizations weren't possible before.