Fixed signed vs. unsigned comparison warning given by MSVC9
I kept getting the warning

  networkoutput.cpp(112) : warning C4018: '<' : signed/unsigned mismatch

This is because writeTo() had a signed return type (int) but data.size()
is an unsigned type. Since writeTo() can never return a negative value,
lets cast the number of bytes which were written to size_t and then return
that to fix the warning.