Deserialize and store process start times
Memorizing the time at which a process was started is necessary to make
the process IDs unique, at least unique for a given host. The process IDs
are reused and thus different processes (with the same name!) might get the
same PID. The only way to tell them apart is by looking at when the
process was started.