Let Trace objects notice when the process is shutting down
A global ShutdownNotifier object can be used to register observers; the
observers will get called once the process shuts down.

Technically, the Trace objects will be called at the point global non-static
objects are destroyed. When a Trace object notices a shutdown, it will send
out a 'ProcessShutdownEvent' if possible.

A ProcessShutdownEvent contains information about what process stopped and
when it stopped. This event can be serialized by all Serializer objects. If
possible, the Trace objects will catch this event and then serialize & write
it out.