Allocate the default Trace object on the heap
This object never gets freed, but this way we ensure that the object lives
long enough to notice that the process is shutting down. In particular, it
must live longer than all global non-static objects.

By allocating the object on the heap, the object exists long enough to get
its handleProcessShutdown() reimplementation called.