Don't require that installCrashHandler() is called manually
For convenience, we setup a global object which installes the crash
handler in its constructor. This is probably sufficient for most cases.

If that's not early enough, then people can still call installCrashHandler()
themselves in some code which is executed even earlier. Calling
installCrashHandler() multiple times will only install the handler once.