Try hard to make backtrace generation work on FreeBSD
This patch makes cmake compile a little sample program which attempts
to use the backtrace() function. The function is available in the GNU
C library, so this should work on Linux.

However, on FreeBSD, this test fails. In that case, try to find
libexecinfo, which provides roughly the same functionality.