Avoid violating an assertion in Backtrace::frame()
I noticed that in case a crash is caught, generating a backtrace fails for
me. This caused an abort() in Backtrace::frame() because we always
assumed that generating a backtrace works in case the application
crashes and hence always accessed Backtrace::frame(0).

Fix this by testing the number of frames for the generated backtrace; in
case the backtrace has no frames, just print dummy file names, line numbers,
function names.