Turned the VariableValue::Type enum into a full VariableType structure
The advantage of promoting the type is that the list of types is easier
to extend (by just modifying variabletypes.def); it's also possible to
iterate over the list of types, and it's easy to get the string
representation for any type.

Unfortunately, I had to change the numbering of types now; 0 is not
'String' anymore but 'Unknown'. String is now '1'. This means that
using the VariableType::valueAsString() function on variable types which
are stored in old databases yields the wrong string.