Made it possible to suspend and resume the EntryItemModel
Also, this patch optimizes the model a bit by not having a QTimer running
all the time but only if an entry was added. Furthermore, we don't keep
a list of new TraceEntry objects around anymore, we just remember how many
there were because that's all the information we need so far.