Fix what I believe to be a thread safety issue.
Inside QThread::run() 'this' is not part of the current thread.
Therefore the connect() statement crosses threads and the readAll()
call on the socket will be performed across boundaries. This is
not safe I assume? Could Frerich's empty 'data' problem.
Helgrind was seeing problems with the QRingBuffer in the socket
at least.

I've run into the 'this' problem with connect() before. Never
found a nicer solution than going through another QObject
created in the thread.