Print out process's start and end time in XML output
Now grouping the process's information within an extra section. Since
pid is a simple attribute like the others (espc. it cannot be used
solely for identifying processes), putting it into a tag like the
others, too.

Considering adding a "real" id (the process table's id) and/or putting
the process information separately to allow improved parsing.