Splitted public API out of tracelib.h
This commits trims down the public API exposed by the trace library. Most
of the code was moved out of tracelib.h into a dedicated trace.h header (which
I wanted to introduce anyway, since the main class is Trace).

As a consequence, we need to ship much less headers.