Merge back the shippedexamples into the examples.
This means less maintenance effort as the same example app code is being
used for building the samples with tracelib as well as having the examples
shipped with a binary tracelib package (or just installed with make
install) for later use.

This requires a small 'dummy' Findtracelib.cmake module to satisfy the
find_package() call and set up the variables used by the examples.