Semi-hacky way to make live searching work
This patch connects the search widget to the entry item model. For
each change in the search, a carefully constructed SQL query is
executed which yields all trace entry IDs which match the specified
search criteria. Those entries are the memorized in a set and then a
global dataChanged() signal is emitted to enforce that the views
update themselves.

The views will then ask the mdel for the new data. The model now
yields a special value (yello) for the Qt::BackgroundRole in case a given
index is contained in the set of entries which matched the search.

The net effect: search results are updated live and the matching entries
get a yellow background.