Use Qt4 imported targets instead of absolute filepaths
This should fix the problem of linking the addressbook example
against tracelib_qtsupport. For static libs CMake always adds all libraries
the static lib was linked against to the list of 'interface libraries' as
all of them are required to be linked against when linking against the
static lib. Hence we ended up with absolute paths to Qt libs in
the generated tracelibTarget-XXX.cmake files. Instead now use imported
Qt targets (as supported by current cmake versions) so we get Qt4::QtCore
instead. This requires the same use of imported Qt targets in the apps
cmake-file, otherwise the imported target is not going to be resolved.