Don't let test programs link against tracelib
Instead of linking against all of tracelib, just compile the (few!)
required source files in. This makes it easier to port the test library
from a static library to a shared library. It's also a nice way to
see the dependencies of the tested code because all dependencies need
to be compiled into the test program.