Added a Server::trimTo() method for trimming the database
The idea of the trimTo() function is that it takes an argument which specifies
how many entries to keep; it will keep the N most recent entries. So far, I
only implemented the 'clear all entries' case (that is, keep zero entries)
since I need this functionality for the GUI. Doing this special case using
dedicated SQL commands makes it very fast. I decided not to do a VACUUM after
clearing the tables because the VACUUM might take quite some time. If people
really want to free up disk space, they have to quit the GUI and just delete
the database file altogether.

I'm undecided whether to put this into the Database class or into the Server.
The point of the Database class is unclear to me, it seems to be a bag of
static utility functions. Put it into the server for now so that we can use
the nice Transaction helper class.