Don't try connecting to server indefinately
This patch is a quick fix to make the AUT (Application Under Trace) not
get bogged down in case a network output was configured but the connection
cannot be established. We used to attempt a connect() call for every visited
trace point, and a failing connect() can take many seconds.

Eventually, it would be nice if we could keep trying to connect() in the
background or so. For now, let's just try connecting once. If it fails, we
memorize that. Customers can enforce that a connection attempt is made again
by touching the tracelib configuration file; doing so will trigger creating
a new Output object.

RESOLVE:APP-40