Allow wildcard as well as strict searching
This patch introduces three radio buttons next to the search field. These
buttons can be used to determine what matching mode is to be used when
searching the trace data. Right now, only strict and wildcard matches
work.

The GUI for regexp matching is there, but making it work requires doing
some calls with the native SQLite API - which we don't have access to
right now.