Relax inprocess library when reading configuration file
This patch changes the configuration reading behaviour of the inprocess
library so that it no longer expects a single filter element below
<tracepointset>. If you wanted to have multiple filters, you needed to
enclose them in a <matchallfilter> or <matchanyfilter> (corresponding to
boolean AND and OR).

This is no longer necessary. Instead, you can now just say

<tracepointset>
  <pathfilter>..</pathfilter>
  <functionfilter>..</functionfilter>
</tracepointset>

and this will behave as if the two filters were contained in a
<matchallfilter>.