Introduced VariableValue::convertToString()
This (static) function takes a VariableValue object and returns the string
representation of the value. This is not the same as the asString() member
method because the latter won't return anything at all for values which
are not of type VariableType::String.

I'll use this function in a new macro, hence it needs to be exported.