Remove Qt from link-interface of hooklib_qtsupport
This is necessary so our binary packages don't
require cmake 2.8.2 or have the Qt paths from the
build machine in them. Instead now when somebody
wants to use hooklib_qtsupport its necessary to
link to Qt explicitly (instead of relying on
the implicit linking from the interface).

Also move the qtsupport lib after the tracelib
declaration as it depends on that.