Make it possible to inspect and configure trace keys via the GUI
This patch adds a new control to the 'Filter' page which shows all the
trace keys stored in the trace file. The items can be checked and unchecked,
which is going to affect the filtering. For now it has no effect.

The texts shown in the GUI are the same keys which are specified using the
_KEY macros in the application. We might want a mapping lateron in case the
internal keys are too cryptic, so that users can assign a more readable
name in the GUI.