Let the server use a SAX parser for reading received XML data
The advantage of using SAX here is that we don't have to fiddle with
finding the start and end position of an XML document by hand. We can
just feed data to the parser incrementally. The old version had bugs in
case a lot of data came in very quickly.