Introduced a customized QHeaderView version to speed things up.
Turned out that QHeaderView::sectionSizeFromContents() got called
for *all* rows which slowed down things tremendously as we display
SQL data. Now we return a fixed size.

An alternative solution we considered was a custom Qt::SizeHintRole
implementation in EntryItemModel::data(). That would make
QHeaderView also stop querying the model.

TODO: calculate the optimal number of digits for the fixed width.