Avoid using the OFFSET clause in SQL queries
I noticed a slowdown in queryForEntries() which came apparent when scrolling
around in larger data sets. The issue was that the SQL statement included the
OFFSET clause which is a bit expensive in that it requires to SELECT all
preceding entries as well (only to skip them).

I believe we can use the fact that trace entry IDs are always increasing
here. Every time the number of matching entries might have changed (because
new entries were added or the filter was changed), compute a row->ID
mapping for *all* entries. We can use this information to replace the OFFSET
clause in the SQL query which actually retrieves data with a
'trace_entry.id >= x' where 'x' is the id for the given start row.

This makes scrolling acceptable over the complete range for me.