Let highlighting entries work on our internal cache
The search which highlights entries became rather slow because the
internal SQL usage was quite inefficient. I didn't bother trying to fix
that but instead decided to just let the search operate on the internal
cache only. The advantage of this is also that we only search among the
previously filtered entries.