Kicked out 'Variables' column from trace entry table again
This column caused quite some problems since it was different from the
other columns in that no single SQL field (hence no single QVariant in
the m_data vector) corresponded to the value of the cells in that column.

This column also caused some performance issues because it needed to do
SQL queries to compute the string value displayed.