Extended storage settings dialog for server settings
This patch extends the Settings class (as well as the storage settings
configuration dialog) with a few fields to influence settings for the
storage server. The GUI offers to start traced automatically (in which case
an output file name has to be given) or it can connect to a running traced
(in which case the port at which the server is listening must be specified).