Capture and display traced output.
Without this it's extremely hard to find out about some errors.
Like me accidentally having a .trace file with an old version lying
around.

Showing output with a message box could be a bit nasty if the
daemon generates some useless noise. It better doesn't. Or we
have to collect all output in some log window...