Make default convertVariable() specializations for primitive types inline
By now it becomes clear that the convertVariable() specializations for
basic C++ types are all very small. We can thus safely make them inline. This
greatly reduces the numbers of symbols which we need to export from
tracelib.dll and it makes it easier to extend this list in the future (since
the list is not duplicated in the header file and in the .cpp file anymore).