Let XML serializer dump all available trace keys
It's a bit wasteful to serialize (and eventually send) the list of all
trace keys again and again. Unfortunately we don't know when we need to send
this though since clients can come and go all the time. So we use the same
approach as we did when sending the process start time: we send it for every
event.