Added handles for Ctrl+C (and similiar) on Windows
I noticed a few cases in which traced kept running even though it shouldn't. It
happened when Ctrl+C'ing out of the process, or when closing the window of the
console.

I first thought of installing a SIGINT handler for this but it seems that
Windows has a dedicated function for handling Ctrl+C and other cases which
cause a process to terminate.

Use this to ensure that the server always closes in a clean manner (writing out
any data we still have pending).