Document lifetime requirements on the TRACELIB_VAR argument
Thomas Lettner recently hit a problem in which he passed a pointer to an
object whose lifetime was shorter than the TRACELIB_WATCH_STREAM macro
invocation.

Since all the arguments to TRACELIB_VAR (or fVar) are only actually
evaluated at the end of the TRACELIB_WATCH_STREAM (or fWatch) statement,
the given values must be valid up to that point, too.

Document this.