Use setSerializer()/setOutput() when re-reading the configuration
Using these functions saves some code, and it makes sure that we cannot forget
locking the respective mutex (and possibly setting the variable to zero, in
case we want to).

In addition to resolving some code duplication, this patch also makes sure
that m_output is null when the configuration file wasn't read.