Fixed GUI to update properly in case trace entries are generated very quickly
The problem was that for each read event on the socket, the GUI would read
only one trace entry. However, if the trace entries arrive quickly, the
server would send multiple entries in one go. So we have to keep track of
the payload size of datagrams and try to read as many entries as possible.