Implement resetting trace entry ID when nuking tace more defensively
Instead of resetting the ID field (including the trace_point_group table,
which isn't even cleared!), let's just reset the ID for the trace_entry
table. This achieves the effect I need and reduces the risk for unintended
side effects. It doesn't hurt if the other tables keep on using old IDs.