Don't let hooklib crash on empty <key> elements
TiXmlElement::GetText() returns a null pointer instead of an empty string
for elements without a text (like <key></key>). Guard against this by using
a helper function for retrieving the text of an element instead of calling
GetText() directly. We actually had the same bug in a few other places, which
get fixed by this as well.

RESOLVE:APP-100