Fixed various problems with nuking by letting traced do it
We experienced all kinds of crashes and locking issues when nuking the
database. It seems the cause for all this trouble was the fact that the
server (traced) writes new entry to the database but the GUI (tracegui)
nukes it. Those two processes can conflict when accessing the database.

Fix this by letting the GUI tell the server that the database should be
nuked. The server will call back as soon as it's done.

RESOLVE:APP-102
RESOLVE:APP-103