Let the tracelib.xml file support a <storage> element
This element is supposed to be used for defining how the trace data should
be stored. A maximum trace file size can be specified as well as a percentage
of the trace entries which should be moved into a separate file as soon as
the maximum a hit.

This could do with some extra error checking for invalid values but in
principle it should do.

RESOLVE:APP-99