Gross hack to transport tracelib.xml configuration to traced
We're hijacking the XML stream of <traceentry> elements to transport the
current storage configuration to the server. This is quite ugly (it introduces
an ugly dependency of the Serializer hierarchy on StorageConfiguration) and
in efficient (the configuration is transmitted for every single entry) but
it's easy, and I'm running out of time.