Removed hard limit/soft limit configuration settings from GUI
These settings are outdated (there is no hard and soft limit anymore but
just a maximum size and a percentage by which to shrink). Furthermore, this
setting is not per server but per tracelib.xml.

Appropriate GUI elements might be reintroduced at some point in the
tracelib.xml configuration editor.