Let EntryItemModel check itself by using the Qt ModelTest
This is a copy of the Qt Model Test class
(http://developer.qt.nokia.com/wiki/Model_Test) which continuously checks a
model as it changes, verifying a proper state. I modified it slightly so that
it doesn't use QTestLib but rather really assert()'s.

Our model is rather buggy and this should help improving the state of things.