Fixed storage element saving issue.
The problem was in missed storage element in configuration file
after saving it from GUI (configurator dialog), if it was in the
file before. The reason was that storage settings weren't read
by configuration file parser and therefore weren't saved on file
saving. With this patch we read, store and save this element in
the file.