Avoid passing std::string across DLL boundaries
std::string allocates and releases memory internally. However, with Windows
DLLs, all allocations/releases of some memory region have to be done in the
same module since the memory manager is per-module.

Fixes a few crashes when trying to use tracelib with an application which
statically links against the C runtime.