Drop manual copying of Qt libs and use DeployQt4 for that
This is a cmake module that transparently copies all the needed Qt libs
to the executable and works across all platforms. Just need to specify we
also need the qsqlite plugin.

The generated qt.conf does not look as if it would work and for linux we
might want to try setting up RPATH/RUNPATH.

Need to call INSTALL_QT4_EXECUTABLE in the same scope as the target is
created, otherwise the generate install-code will try to fixup not-yet
installed files.