Fix link interface
The link-interface is supposed to indicate which libraries are re-exposed
in the targets ABI. So for example if hooklib would expose structs or
other symbols from pcre or other libs it links against in its public ABI
then the those libraries would need to be in the link-interface.

However the ABI of tracelib is specifically small and only exposes the
tracelib symbols and STL symbols.