Drop the MSVC-suffix
Its not necessary anymore since we link statically against the C runtime
and do not delete objects passed into the library from the calling side.
So the tracelib can be built with any compiler and the user with a totally
different one without causing any problems (at least in theory).