Make sure that VariableValue allocates and deallocates in tracelib
Now that tracelib is statically linked against the C runtime, we have
to be careful where memory is allocated and deallocated. In the case at
hand, a call to VariableValue::stringValue("foo") (which is not inline!)
would result in the std::string member variable to be constructed inside
tracelib, i.e. the memory is also allocated inside tracelib.

Since no destructor was specified, the compiler synthesized a public
*inline* destructor, i.e. the destruction of the std::string and all
memory deallocation calls associated with that happened on the caller
side. That's where we had a mixup between allocation/deallocation.

This fixes at least one crash I observed when doing something like

  TRACELIB_TRACE_MSG("aaaaaaaaaaaaaaaaaa")