Support componentized packages of tracelib
This can generate 2-3 packages, one for the library, the docs and the
applications or alternatively one for the runtime-stuff (traced, gui etc.)
and one for the hooklib.

The idea here is to be able to drop the 'hooklib_only' hack but still keep
a way to simply install just the hooklib library on some machine - or just
the runtime apps for analyzation.