Enforce usage of a static libbfd
This makes tracelib more portable since the dynamic bfd library has the
version encoded into the filename and hence tracelib only works on systems
having the exact same bfd library version installed.

The fact that this is the case may hint that using newer bfd libraries on
older code or vice versa could be a problem, but hopefully the bits we do
use are not affected by that.