Fix casting of pthread_t with older gcc's on Linux
gcc 4.3 does not like reinterpret_cast'ing from an unsigned long to
an unsigned long (newer gcc's don't have this problem), but we do need a
cast for gcc 4.2.1 on MacOSX where pthread_t is actually a pointer to a
struct.

This is not technically correct, pthread_t is an opaque type and hence
casting it to a number can fail (if its a struct). However there should be
only very few, seldomly used platforms that actually run into that, so
adding a 'proper' solution is not worth it at the moment.