Make all executables app bundles on MacOSX
Unfortunately the DeployQt4 module we use to include Qt in the
installation directory (to make it redistributable) does not support
non-app-bundle executables on MacOSX. Fixing it up is not very easy
since the underlying BundleUtilities module also seems to make some
assumptions that do not hold for our installation type.

So making these apps app-bundles is the 'easy way out' on the expense of
making their use a bit more cumbersome.