Improved support for MSVC6
We don't need to build tracelib itself with MSVC6, but it would be nice if the
header files can be processed by MSVC6 so that clients don't have to use a more
recent compiler.

First step is to look for a replacement for __FUNCSIG__ - I couldn't find
anything at all, so for now let's just say "unknown" (i.e. we cannot figure out
what the currently executed function is).