Flush the handle-oriented outputs after each line
With lots of data to be produced this is important since otherwise
significant amounts of data may be lost.

This is has a runtime overhead since flushing takes time, so we may need
to add an option to disable this. Ultimately though if performance is
crucial the best option is to use the trace daemon.