Make tracelib compile against libiberty on newer Debian systems
I've noticed this while building it on my unstable system, apparently
demangle.h and the corresponding library containing the symbols have been
split off into a separate library. Hence if the header is not found
(in /usr/include usually) try to find libiberty include directory and
library and link against that for the demangle stuff.