Silence warning from CMP0022 policy
the LINK_INTERFACE_LIBRARIES stuff is deprecated and one should use the
LINK_PUBLIC/LINK_PRIVATE flags to indicate which dependencies should be
're-exported' and which ones should be considered private.

So this change should have no difference on the visible dependencies of
libtracelib.so but ensures cmake does not produce the nasty warning.