Fix usage of the longer macros like TRACELIB_TRACE_MSG
Turns out my earlier replacement of msg with convertVariable(msg) was a
bit overzealous. The msg in these longer macros is actually expected to
already be something like '"info" << fValue(variable)' and hence wrapping
that in convertVariable will generate compile errors since the left side
is a string constant which has no << defined.