Proper fix for the global operator<< and Stringbuilder
The issue is that the global operator<< takes a reference and returns one,
which is needed since it changes the object it is given. At the same time
the calls so far used a temporary unnamed object which is seemingly not
allowed by all compilers (specifically gcc/clang).

So the object needs to get a real name, but this requires quite some
changes as the non-message variants should produce a 0 as char* and not
an empty string (there's code that relies on this). Hence need to duplicate
the macros into two versions one which builds a message and one without.

Tried to share the common code via yet another set of macros (ugly, but
probably better than having to change things around in 2 places).