Drop the convoluted _BEGIN/_END macros again
After reading the code with a fresh mindset and with the common parts moved
out it became evident that there is a way to let the caller provide a
code-snippet that evaluates to either a StringBuilder variable or a
null-pointer-to-char.

So use that as it makes the code a bit more readable.