Support milliseconds in timestamps.
Summary:
These changes enable tracer to generate timestamps in milliseconds since unix
epoch and print them in the serializers as well as showing the information in the
gui.

A change in the database format is required, but convertdb can be used to
upgrade the DB without loosing any data.

The only real issue would be trace2xml which now produces timestamps in milliseconds
since unix epoch instead of the textual format from the sqlite database.

Reviewers: frerich

Reviewed By: frerich

Differential Revision: http://phabricator.froglogic.com/D24