Fix order of trace statements within a scope
The changes to give the TracePointVisitor a name (to support msvc6)
generates code which forces the trace messages in a given scope to be
serialized in the inverse order of their creation in the code. The reason
is that local objects are being destructed in the reverse order than they
are constructed and the TracePointVisitor uses its destructor to 'flush'
its data.

Avoid that by using a small proxy class which allows to detect the last
stream operation and then flushes the TracePointVisitor it manages. This
way the flush happens at the next semicolon after the last streaming operator
is called.