Fixed watch points in Windows builds
A crash occurred when the std::vector (which is allocated on the caller
side) got elements appended, because that made the vector allocate
memory - but this happened in the CRT, i.e. it was not inline. Hence,
when the vector was *deleted* (which also happened on the caller) site,
an inline destructore was invoked which tried to delete[] the internal
memory of the vector. So the new[] happened in the CRT and the delete[]
in the caller code, which broke.

Fixed this by not making VariableSnapshot a typedef for std::vector;
instead, make it a proper class which has non-inline constructor,
destructor and insertion operations so that all std::vector operations
happen inside tracelib.