Implement a xml2trace tool
This generates a sqlite database that can be loaded into the GUI from a
xml file. In particular useful to let a third-party generate trace logs
in xml format (easier to setup than the traced) and then receive them and
analyze them in the GUI.

Unfortunately does not quite work with either the xml generated by
hooklib directly nor by the xml the trace2xml tool generates.