Make the provided TRACELIB_INCLUDE_DIR variable more 'standard'
Supporting systems where the lib-install-dir is completely outside the
prefix of the include-install-dir (i.e. the parent dir is not the same) is
a nice idea but causes issues with relocatability. Since such systems are
extremely rare in practice IMO (Linux distro's don't do this and local
installations are also easier to setup specifying just a single path) lets
make this work for the standard case more easily

Also dropped the 'tracelib' directory since the main user (inside the
company) uses #include <tracelib/tracelib.h> already a lot. IMO thats nicer
anyway since it serves as a kind of namespace.