Introduce fixed-size variable conversion for integers
Instead of relying the not-that-well-defined types short, long, int etc.
explicitly define the sizes of integer types we support to stream.

This drops the qlonglong and qulonglong support, since those two are only
available for Qt4 and are typedefs for the 64bit integer type.

In Squish we're having an issue with code that streams size_t which is
a 64bit unsigned integer on 64bit Microsoft compilers (msvc8 in particular)
and building against Qt3. In that case there is no convertVariable for
the size_t (since the qulonglong overload exists only for Qt4) and thus
linking Squish code would fail.

In order to avoid such hard-to-debug problems, we'll just make sure that all
standard sizes are covered for now.

In addition to that this includes a new flag to ensure the signed numbers
are properly printed including the sign. Without this signed numbers would
end up being printed as unsigned number, i.e. tracelib would never print
negative numbers.