Drop char version of convertVariable
Not sure this is actually correct. It seems with current gcc char is
different from int8_t and in fact we'd like to stream char (and maybe also
the already removed signed-char) as strings and not as numbers.

No idea how to distinguish that if the compiler maps 8-bit-int and char
to the same underlying type (char for msvc9 at least). I guess we simply
can't do anything about this.