Improve macro to decide wether stdint.h is used or not
This should improve the safety of the decision to use stdint.h and the
typedefs inside that vs. defining them. Essentially we setup our own typedefs
for MSVC which has no C99 support (and hence lacks the int<bitness>_t types)
and use stdint.h for everything else.

In case we find a compiler with C99 support disabled other than MSVC its
considered to be unsupported since we don't know how to get at the needed
types (in particular 64bit integers) for that case.