Add a simple python script for building tracer
This moves the platform-specific knowledge needed to build tracer, run
the tests and eventually package it into a python script. The python script
can be easily triggered with 'cross-platform' shell invocations in CI systems
and hence can be used to prepare jobs that run the build across different
platforms without the need for configuration-duplication.