Restore arch argument
The jenkins matrix config is easier to setup when including the architecture
as an axis and hence we should pass this to the cibuild to avoid building
things multiple times. So include arch again as argument and build only
the architecture. Only for MacOSX do we support 'universal' as a special
architecture to indicate that both 32 and 64bit should be included.