Prepend the Qt directory instead of appending.
I noticed that CMake is in PATH (intentionally) on the machine where the
test crashes and then realized that CMake comes with its own Qt install
these days. Its not unlikely this is causing issues, so until we use a
renamed Qt on windows, lets prepend Qt in PATH to make sure the right one
is used for running the tests - in a last desperate attempt to avoid the
crash.