Fix packaging on Windows
gettempdir() may return a path inside a users home dir (usually
AppData/Local/Temp) and if the users name is longer than 8 characters this
results in a short-form path name (i.e. froglo~1 instead of froglogic).

Using that as packaging directory triggers an oversight in CMake's bundle
utilities (specificically GetPrerequisites.cmake's gp_resolved_file_type)
where the absolute path of a file is compared to another path obtained
through CMake API. The CMake API will however yield the long form of a path
like the above and since the comparison is done using strequal the check
fails. Hence dll's lying at the same place as the AUT will not be detected
as 'local' and thus will be reported as missing requirements. This finally
aborts packaging as the bundling fails (all this is part of a verification
step that bundling worked).

So in order to be able to build tracelib packages with cmake versions that
do not contain a fix for this, lets convert the packaging dir to a long
pathname ourselves.