Fix calling just fTrace
This fixes the streaming macro when there's nothing streamed into it. In
that case we should still visit the tracepoint, but thats only done when
a proxy is flushed. In this case there's no proxy, so the TPVisitor itself
also needs to verify wether something was streamed into it and otherwise
flush on destruction.