Merge WatchPointVisitor into TracePointVisitor
The main reason is to avoid having to do the same proxy-trick and non-member
templates for WatchPointVisitor to support MSVC6. The difference between
the two visitors was just that WatchPointVisitor did not accept plain
strings streamed into it and TracePointVisitor did not accept variables for
snapshotting.

This change means that fTrace can be used for snapshots just like fWatch,
though this could be a bit surprising when reading the logs.

There's a small catch which is that the global shift-operator's need to
be defined in the right order for the compiler (at least clang) to pick
the AbstractVariable* one where needed. Maybe someone sees a way to do
this in a less fragile way.