Set a timeout for the test execution
This prevents hanging the build in case one of the tests hangs. In particular
useful when running these on windows where a user interaction may be needed
if one of the dll's is missing from PATH.

This is more fine-grained then to setup a timeout in jenkins as it only
affects the problematic test execution and not the building part. In
addition it prevents stalling even outside jenkins.