Support a kind of 'fat' tracelib libraries on non-mac platforms
This renames the shared library to include the bitness into the library
name and thus allows to put 32bit and 64bit tracelib libraries in a single
directory and different apps use either of them as they need.

The packages generated by cibuild.py will now have this renamed library
so we can use that for Squish builds which sometimes have components in
different architectures (runner vs. hook).