Add some more needed overloads of convertVariable for MSVC
Turns out that MSVC <= 10 (can't test higher versions) have long and
unsigned long as separate types (not the same as __int32 or __int64). So
need overloads for those basic types.

In addition MSVC10 has separate types for char and signed char, hence needing
an overload for that case as well.