Make StreamLogOutput not take ownership of the stream
Using a new output type instead which keeps ownership of streams opened
explicitly for files. This allows to re-use StreamLogOutput for standard
streams like std::cerr and is more in line with the usual usage of streams
in C++.

Allow to re-use the stream from a FileLogOutput to be able to have error
and status go to the same file as there's only a single envvar supported
at the moment.