Make MacOSX pick up tracelib.xml properly
progname only yields a name without any slashes, hence the 'use cwd to
absolutify' codepath wasn't use and the app is not in PATH. This led
traeclib to try /tracelib.xml which is not a good idea.

Use MacOSX API to get the actual absolute path to the executable and
then resolve to a real path.