Add CXX/CC to existing environment
It seems like the python I'm using and the one on etch64 have different
semantics for subprocess' env argument than the one on lenny32. On the latter
machine the subprocess' environment is replaced with the values from the
argument while on my system and etch64 the values are added to the current
env automatically.

Fix that oddity by simply doing the extension ourselves so cmake can find
tools like 'make' in the PATH it gets again.