Revert the intxy_t stuff and use short/int/long/long long for integrals
This originally worked quite well, except for a case of unsigned
long or so on some platforms. So instead of getting the intxy_t right for
all compilers, lets just go back to that with the fixes done meanwhile
(using a full 64bit integer for storing the data and storing the sign).